pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "imagekitio>=5.2.0",
    "uvicorn[standard]>=0.42.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]
